Handle multi-line lambdas better
We were running into a problem where if we strip out the opening
bracket we would never find a valid substring that parsed correctly.

We solve this by mangling the lambda onto a single line before we
try to parse it. This also requires dealing with the case where there
is a line continuation character instead of a bracket.

This also improves some handling of cases where our initially
extracted source may be invalid.

Reminder that this function is terrible.